pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "generic-llm-api-client"
version = "0.3.2"
description = "A unified, provider-agnostic Python client for multiple LLM APIs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Sorin Marti", email = "sorin.marti@gmail.com"}
]
maintainers = [
    {name = "Sorin Marti", email = "sorin.marti@gmail.com"}
]
keywords = [
    "llm",
    "ai",
    "openai",
    "anthropic",
    "claude",
    "gemini",
    "mistral",
    "deepseek",
    "qwen",
    "cohere",
    "api-client",
    "humanities",
    "research",
    "benchmarking"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "anthropic~=0.79.0",
    "openai~=2.21.0",
    "mistralai~=1.12.2",
    "google-genai~=1.63.0",
    "cohere~=5.20.5",
    "requests~=2.32.5",
    "geonamescache>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/RISE-UNIBAS/generic_llm_api_client"
Documentation = "https://github.com/RISE-UNIBAS/generic_llm_api_client#readme"
Repository = "https://github.com/RISE-UNIBAS/generic_llm_api_client"
Issues = "https://github.com/RISE-UNIBAS/generic_llm_api_client/issues"

[tool.setuptools]
packages = ["ai_client"]

[tool.setuptools.package-data]
ai_client = ["py.typed", "pricing.json", "tools/default_tools.json"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
